+++ examples/cells-by-row.js
@@ -19,16 +19,27 @@ CellsByRow.prototype._resetLayout = function() {
   this._getMeasurement( 'columnWidth', 'outerWidth' );
   this._getMeasurement( 'rowHeight', 'outerHeight' );
 
-  this.cols = Math.floor( this.size.innerWidth / this.columnWidth );
-  this.cols = Math.max( this.cols, 1 );
+  if ( this.options.isHorizontal ) {
+    this.rows = Math.floor( this.size.innerHeight / this.rowHeight );
+    this.rows = Math.max( this.rows, 1 );
+  } else {
+    this.cols = Math.floor( this.size.innerWidth / this.columnWidth );
+    this.cols = Math.max( this.cols, 1 );
+  }
 
   this.itemIndex = 0;
 };
 
 CellsByRow.prototype._getItemLayoutPosition = function( item ) {
   item.getSize();
-  var column = this.itemIndex % this.cols;
-  var row = Math.floor( this.itemIndex / this.cols );
+  var column, row;
+  if ( this.options.isHorizontal ) {
+    row = this.itemIndex % this.rows;
+    column = Math.floor( this.itemIndex / this.rows );
+  } else {
+    column = this.itemIndex % this.cols;
+    row = Math.floor( this.itemIndex / this.cols );
+  }
   var x = column * this.columnWidth;
   var y = row * this.rowHeight;
   this.itemIndex++;
@@ -39,9 +50,15 @@ CellsByRow.prototype._getItemLayoutPosition = function( item ) {
 };
 
 CellsByRow.prototype._getContainerSize = function() {
-  return {
-    height: Math.ceil( this.itemIndex / this.cols ) * this.rowHeight
-  };
+  if ( this.options.isHorizontal ) {
+    return {
+      width: Math.ceil( this.itemIndex / this.rows ) * this.columnWidth
+    };
+  } else {
+    return {
+      height: Math.ceil( this.itemIndex / this.cols ) * this.rowHeight
+    };
+  }
 };
 
 return CellsByRow;